refactor: improve file hierarchy. refactor: remove dead code
