+++ browser/UI/failure.js
