+++ browser/UI/recordingIndicator.js
