+++ browser/container/NavConsole.js
