+++ browser/index.js
@@ -11,7 +11,7 @@ import Menu from './menu'
 import Navigator from './navigator'
 import Intro from './intro.jsx'
 import introText from './introText.js'
-import {mediaRecorder, startRecording, stopRecording} from './audio'
+import {mediaRecorder, startRecording, stopRecording} from './audio/audio'
 import loadAllAssets from './assets'
 import FailureView from './failureView'
 import { startSyncingPhaseAndStrikes } from './firebase'
