+++ browser/navigator.js
@@ -1,24 +1,22 @@
 import React from 'react'
 import {connect} from 'react-redux'
-//import Clock from 'react-countdown-clock'
 import NavConsole from './container/NavConsole.jsx'
-import {setUpRecording} from './audio' 
+import {setUpRecording} from '.audio/audio'
 
-export default class Navigator extends React.Component{
-  constructor(props){
+export default class Navigator extends React.Component {
+  constructor(props) {
     super(props)
   }
 
-  componentDidMount(){
+  componentDidMount() {
     setUpRecording(this.props.isNavigator)
   }
 
-  render(){
+  render() {
     return (
     <div>
       <NavConsole/>
     </div>
     )
   }
-
 }
