+++ browser/simulation.js
@@ -15,7 +15,7 @@ import _ from 'lodash'
 import {solution1, solution2, solution3} from './validation'
 import {SpaceshipAmbience} from './soundEffects'
 import Failure from './failure'
-import {setUpRecording} from './audio'
+import {setUpRecording} from './audio/audio'
 import {setButtonPressedColor, resetButtonPressedColors, resetClickHandlers} from './UI'
 import stopDefaultAndPropagation from './utils/events'
 import {MODULE_ONE, MODULE_TWO} from './utils/constants'
